Added: Data module. Extracted data manipulation functions
