+++ src/cli/mod.rs
@@ -1,4 +1,4 @@
-use clap::{Arg, App, ArgMatches};
+use clap::{App, Arg, ArgMatches};
 
 pub fn get_cli_helper<'a>() -> ArgMatches<'a> {
     App::new(env!("CARGO_PKG_NAME"))
@@ -11,7 +11,7 @@ pub fn get_cli_helper<'a>() -> ArgMatches<'a> {
                 .value_name("in")
                 .help("The csv file to read")
                 .takes_value(true)
-                .required(true)
+                .required(true),
         )
         .arg(
             Arg::with_name("dimensional-separator")
@@ -19,7 +19,7 @@ pub fn get_cli_helper<'a>() -> ArgMatches<'a> {
                 .long("dimensional-separator")
                 .value_name("dimensional-separator")
                 .help("A separator to break header names allowing you to create deeper objects")
-                .takes_value(true)
+                .takes_value(true),
         )
         .get_matches()
 }
